requirements.txt
streamlit==1.35.0
pandas==2.2.2
langdetect==1.0.9
scikit-learn==1.4.2
transformers==4.41.1
torch==2.3.0
sentence-transformers==2.7.0
numpy==1.26.4
langchain>=0.1.16
langchain-community>=0.0.26
langchain-google-genai>=0.0.6
langchain-huggingface>=0.0.6
faiss-cpu==1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.1
langdetect>=1.0.9
transformers>=4.40.1
torch>=2.1.0
scipy>=1.10.1
